+++ lindera/src/token_filter/japanese_katakana_stem.rs
@@ -30,6 +30,7 @@ impl JapaneseKatakanaStemTokenFilterConfig {
     }
 }
 
+#[derive(Clone, Debug)]
 pub struct JapaneseKatakanaStemTokenFilter {
     config: JapaneseKatakanaStemTokenFilterConfig,
 }