+++ lindera/src/token_filter/length.rs
@@ -22,6 +22,7 @@ impl LengthTokenFilterConfig {
     }
 }
 
+#[derive(Clone, Debug)]
 pub struct LengthTokenFilter {
     config: LengthTokenFilterConfig,
 }